chore: move tests alongside source files in pkg
