+++ examples/basic/main_test.go
@@ -1,4 +1,4 @@
-package main
+package main_test
 
 import (
 	"context"
